* generic/tclProc.c: corrected error reporting for default case 	at the global level for uplevel command.
	* generic/tclIOSock.c: changed int to size_t type for len
	in TclSockMinimumBuffers.

	* generic/tclCkalloc.c: fixed Tcl_DbCkfree to return a value
	on NULL input. [Bug: 3400]

	* generic/tclStringObj.c: fixed support for passing in negative
	length to Tcl_SetUnicodeObj, et al handling routines. [Bug: 3380]

	* doc/scan.n:
	* tests/scan.test:
	* generic/tclScan.c: finished support for inline scan by
	supporting XPG identifiers.

	* doc/http.n:
	* library/http2.1/http.tcl: added register and unregister
	commands to http:: package (better support for tls/SSL),
	as well as -type argument to http::geturl. [RFE: 2617]

	* generic/tclBasic.c: removed extra decr of numLevels in
	Tcl_EvalObjEx that could cause seg fault. (mjansen@wendt.de)

	* generic/tclEvent.c: fixed possible lack of MutexUnlock in
	Tcl_DeleteExitHandler [Bug: 3545]

+++ generic/tclIOSock.c
@@ -8,7 +8,7 @@
  * See the file "license.terms" for information on usage and redistribution
  * of this file, and for a DISCLAIMER OF ALL WARRANTIES.
  *
- * RCS: @(#) $Id: tclIOSock.c,v 1.3 1999/04/16 00:46:47 stanton Exp $
+ * RCS: @(#) $Id: tclIOSock.c,v 1.4 1999/11/19 06:34:23 hobbs Exp $
  */
 
 #include "tclInt.h"
@@ -91,8 +91,8 @@ TclSockMinimumBuffers(sock, size)
     int size;			/* Minimum buffer size */
 {
     int current;
-    int len;
-    
+    size_t len;
+
     len = sizeof(int);
     getsockopt(sock, SOL_SOCKET, SO_SNDBUF, (char *)&current, &len);
     if (current < size) {
